feature: add at-sample e2e test

+++ e2e-test/e2e-test-builder/src/main/java/org/apache/seata/builder/ImageBuilder.java
@@ -66,7 +66,9 @@ public class ImageBuilder {
                 }
             }
         }
-        Files.delete(dir);
+        if (dir.toFile().exists()) {
+            Files.delete(dir);
+        }
     }
 
     private void replaceFilesInModule(File moduleDir, E2EConfig e2EConfig) throws IOException {
